requirements.txt
numpy>=1.24
Pillow>=10.0
pytesseract>=0.3.10
matplotlib>=3.7
